restructuring use effect calls
