+++ src/app/context/AccountContext.tsx
@@ -119,7 +119,6 @@ export const AccountProvider: React.FC<{ children: React.ReactNode }> = ({ child
 
   const addAccount = async (account) => {
     const {transactions, ...accountWithoutTransactions} = account;
-    console.log('account', account)
     const { data, error } = await supabase.from("accounts").insert([
       {
         ...accountWithoutTransactions,
@@ -132,8 +131,6 @@ export const AccountProvider: React.FC<{ children: React.ReactNode }> = ({ child
     if (error) {
       console.error("Add account failed:", error);
     } else {
-      console.log('account', account)
-      console.log('data', data)
       addTransaction(data[0].id, account.transactions[0])
       setAccounts((prev) => [...prev, { ...account, id: data[0].id}]);
     }
